# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comt@cts Inc API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  avatarInitials: string;
  title: string;
}

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus];

export const ContactStatus = {
  online: "online",
  away: "away",
  offline: "offline",
} as const;

export interface Contact {
  id: number;
  name: string;
  email: string;
  phone?: string;
  company?: string;
  title?: string;
  avatarInitials: string;
  status: ContactStatus;
}

export interface CreateContactInput {
  name: string;
  email: string;
  phone?: string;
  company?: string;
  title?: string;
}

export type ConversationContactStatus =
  (typeof ConversationContactStatus)[keyof typeof ConversationContactStatus];

export const ConversationContactStatus = {
  online: "online",
  away: "away",
  offline: "offline",
} as const;

export interface Conversation {
  id: number;
  contactId: number;
  contactName: string;
  contactAvatarInitials: string;
  contactStatus: ConversationContactStatus;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
}

export interface CreateConversationInput {
  contactId: number;
}

export interface Channel {
  id: number;
  name: string;
  description?: string;
  memberCount: number;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
}

export interface CreateChannelInput {
  name: string;
  description?: string;
}

export interface Message {
  id: number;
  body: string;
  senderName: string;
  senderInitials: string;
  isMine: boolean;
  createdAt: string;
}

export interface SendMessageInput {
  body: string;
}

export interface DeleteResult {
  success: boolean;
}