# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Comt@cts Inc API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contacts
    description: Business contacts
  - name: conversations
    description: Direct message conversations
  - name: channels
    description: Team group channels
  - name: me
    description: Current user
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: List all contacts
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{contactId}:
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /conversations:
    get:
      operationId: getConversations
      tags: [conversations]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}/messages:
    get:
      operationId: getConversationMessages
      tags: [conversations]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendConversationMessage
      tags: [conversations]
      summary: Send a message to a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /channels:
    get:
      operationId: getChannels
      tags: [channels]
      summary: List all channels
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"

  /channels/{channelId}/messages:
    get:
      operationId: getChannelMessages
      tags: [channels]
      summary: Get messages in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendChannelMessage
      tags: [channels]
      summary: Send a message to a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarInitials:
          type: string
        title:
          type: string
      required:
        - id
        - name
        - email
        - avatarInitials
        - title

    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        title:
          type: string
        avatarInitials:
          type: string
        status:
          type: string
          enum: [online, away, offline]
      required:
        - id
        - name
        - email
        - avatarInitials
        - status

    CreateContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        title:
          type: string
      required:
        - name
        - email

    Conversation:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        contactName:
          type: string
        contactAvatarInitials:
          type: string
        contactStatus:
          type: string
          enum: [online, away, offline]
        lastMessage:
          type: string
        lastMessageAt:
          type: string
        unreadCount:
          type: integer
      required:
        - id
        - contactId
        - contactName
        - contactAvatarInitials
        - contactStatus
        - unreadCount

    CreateConversationInput:
      type: object
      properties:
        contactId:
          type: integer
      required:
        - contactId

    Channel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        memberCount:
          type: integer
        lastMessage:
          type: string
        lastMessageAt:
          type: string
        unreadCount:
          type: integer
      required:
        - id
        - name
        - memberCount
        - unreadCount

    CreateChannelInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name

    Message:
      type: object
      properties:
        id:
          type: integer
        body:
          type: string
        senderName:
          type: string
        senderInitials:
          type: string
        isMine:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - body
        - senderName
        - senderInitials
        - isMine
        - createdAt

    SendMessageInput:
      type: object
      properties:
        body:
          type: string
      required:
        - body

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success