# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comt@cts Inc API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  name: string;
  description?: string;
  memberCount: number;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
}