# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comt@cts Inc API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactStatus } from "./contactStatus";

export interface Contact {
  id: number;
  name: string;
  email: string;
  phone?: string;
  company?: string;
  title?: string;
  avatarInitials: string;
  status: ContactStatus;
}