# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comt@cts Inc API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  body: string;
  senderName: string;
  senderInitials: string;
  isMine: boolean;
  createdAt: string;
}